refactor: refactor directory structure for food card components
