+++ src/features/foods/components/index.ts
@@ -1,6 +1,6 @@
 export { AddButton } from './AddButton';
 export { BadgeList } from './BadgeList';
 export { FilterButton } from './FilterButton';
-export { FoodList } from './FoodList';
+export { FoodCardList } from './FoodCardList/FoodCardList';
 export { FoodSort } from './FoodSort';
 export { SearchBar } from './SearchBar';
